✨ first version of the library
I didn't write this first version of the library, [rdubois-crypto](https://github.com/rdubois-crypto)
did. This first implementation comes from the repository linked in the footer,
the repository contains differents crypto implementation, take a look to them.
My role for this repository is to transform Renaud's implementation as a
library easy to use, easy to install, easy to collaborate to, safe and documented.
I also imported the relevant tests. Those tests use the cheatcode `readFile`
to read a JSON file generated by wychproof (a tool to generate test vectors)
and the cheatcode `ffi`. The `ffi` cheatcode is used to execute a sage precomputation
script. That means for the moment you need to install sage to run those tests.
https://github.com/rdubois-crypto/FreshCryptoLib

Co-authored-by: rdubois-crypto <renaud.dubois@ledger.fr>