♻️ split secp256r1 and ECDSA functions
Split the logic of the secp256r1 curve and the ECDSA functions into two
separate files. The secp256r1 file contains the curve parameters, some
useful constants and the free functions for the curve. The ECDSA file consumes
the secp256r1 file and implements the ECDSA related functions.
The ECDSA library use the using..for directive in order to improve the
readability of the code by overriding the uint256 value type.