✨🚧 create the internal precompute variant
This variant perform better than the external precompute variant, but it
requires a more complex off-chain workflow. Indeed, you will need to generate
the precomputed table off-chain too but this time, the precompute table
must be live in the bytecode of the contract that call this library. This
implementation will then use the `codecopy` opcode (instead of the `extcodecopy`)
to copy the precompute table from the contract bytecode to the memory.
It's a cheaper alternative.

❗ This implementation is not yet tested. It is considered higly experimental.