♻️ `ECDSA256r1PrecomputInternal` minor rework
As stated in the file, this implementation is still a work in progress and
must be considered as such. The current implementation is not yet tested,
that's why I refrain myself from revamp it to use the new `U256Modp` type.
In this commit, I change the signature of the `verify` function to match
the signatures of the other implementations, I removed the assembly block
in the verify signature and I use the `pModInv` function in the `mulmuladd`
function.

⚠️ Please note that this update brings a significant alteration - the
signature of the verify function has been modified.