🔧 adopt the naming convention for tests
By adopting the convention, we blacklist non-relevant tests from the gas
report. That produces a more realistic and accurate report.
In addition, I decide to use the suffix `_ReportSkip` to blacklist tests
that do not expect to revert but are not relevant for the gas report
because they test a failure case. Helpful and relevant in a case of a
library that returns a boolean that is supposed to return true when
correctly used.