⚰️ Remove implementation alternatives
So far, we've continued to maintain the default implementation and its
two alternatives. As we don't intend to use either of them for the time being,
we've decided to temporarily remove them from the repository.

If you are looking for the implementation of the alternatives, a frozen
branch named `version/precomputes` is available to fetch from the remote.